+++ src/plugins/blocksmith/components/imgur-post-block-modal/getImgurPostBlockId.js
@@ -1,31 +1,39 @@
 const getImgurPostBlockId = (str) => {
-  const IMGUR_POST_BLOCK_ID_REGEX = /data-id="(a|gallery)?\/?\w+/g;
-  const IMGUR_POST_BLOCK_URL = /\/\/imgur\.com\/?(|a|gallery)+\/\w+/g;
-  const hasDataAttr = IMGUR_POST_BLOCK_ID_REGEX.test(str);
+  const IMGUR_POST_BLOCK_HAS_TYPE = /\/(a|gallery)+\//g;
+  const IMGUR_POST_BLOCK_HAS_EMBED_REGEX = /data-id="(a|gallery)?\/?\w+/g;
+  const IMGUR_POST_BLOCK_URL_REGEX = /http?s:\/\/imgur\.com\/(a|gallery)?(\w+)(\/?)\w+(\/?)/g;
 
-  if (hasDataAttr) {
-    return (str && str.match(IMGUR_POST_BLOCK_ID_REGEX))
-      ? str.match(IMGUR_POST_BLOCK_ID_REGEX)[0].split('"')[1]
-      : str;
+  // handle embed
+  if (IMGUR_POST_BLOCK_HAS_EMBED_REGEX.test(str)) {
+    return str
+      .match(IMGUR_POST_BLOCK_HAS_EMBED_REGEX)[0]
+      .split('"')[1];
   }
 
-  const result = (str && str.match(IMGUR_POST_BLOCK_URL))
-    ? str.match(IMGUR_POST_BLOCK_URL)[0]
-      .split('/')
-      .filter((s) => {
-        if (s === '' || s === 'imgur.com') {
-          return false;
-        }
+  // handle url
+  if (IMGUR_POST_BLOCK_URL_REGEX.test(str)) {
+    if (IMGUR_POST_BLOCK_HAS_TYPE.test(str)) {
+      const url = str
+        .match(IMGUR_POST_BLOCK_URL_REGEX)[0]
+        .split('/')
+        .filter((s) => {
+          if (!s || s === 'https:' || s === 'imgur.com') {
+            return false;
+          }
+          return true;
+        });
 
-        return true;
-      })
-    : '';
+      return `a/${url[1]}`;
+    }
 
-  if (result.length === 2) {
-    return `a/${result[1]}`;
+    const url = str
+      .match(IMGUR_POST_BLOCK_URL_REGEX)[0]
+      .split('/');
+
+    return url[url.length - 1];
   }
 
-  return result ? result[0] : '';
+  return str;
 };
 
 export default getImgurPostBlockId;